Begin port to multiplatform
