--- lib/build.gradle
@@ -1,51 +1,0 @@
-import cash.z.ecc.android.Deps
-
-plugins {
-    id "org.jetbrains.kotlin.jvm"
-    id 'org.jetbrains.dokka' version '0.10.1'
-    id 'java-library'
-    id 'com.vanniktech.maven.publish' version '0.14.2'
-}
-
-group = project.property('GROUP')
-version = project.property('VERSION_NAME')
-
-tasks {
-    compileKotlin {
-        kotlinOptions { jvmTarget = 1.8 }
-        sourceCompatibility = 1.8
-    }
-    compileTestKotlin {
-        kotlinOptions { jvmTarget = 1.8 }
-        sourceCompatibility = 1.8
-    }
-
-    // Generate Kotlin/Java documentation from sources.
-    dokka {
-        outputFormat = "html"
-    }
-}
-dependencies {
-    implementation Deps.Kotlin.STDLIB
-
-    // Tests
-    testImplementation Deps.Kotest.RUNNER
-    testImplementation Deps.Kotest.ASSERTIONS
-    testImplementation Deps.Kotest.PROPERTY
-    testImplementation Deps.SquareMoshi.MOSHI
-    testImplementation Deps.SquareMoshi.MOSHI_KOTLIN
-    testImplementation Deps.SquareOkio.OKIO
-}
-test {
-    useJUnitPlatform()
-    testLogging {
-        showExceptions = true
-        showStackTraces = true
-        exceptionFormat = "full"
-        events = ["failed", "skipped"]
-        showStandardStreams = true
-    }
-}
-
-sourceCompatibility = "8"
-targetCompatibility = "8"
