+++ lib/src/commonMain/kotlin/cash/z/ecc/android/crypto/FallbackProvider.kt
@@ -1,0 +1,3 @@
+package cash.z.ecc.android.crypto
+
+expect class FallbackProvider()
