+++ lib/src/commonMain/kotlin/cash/z/ecc/android/crypto/SecretKeyCommon.kt
@@ -1,0 +1,9 @@
+package cash.z.ecc.android.crypto
+
+expect class SecretKeyCommon {
+    val encoded: ByteArray
+
+}
+
+
+
