+++ lib/src/commonMain/kotlin/cash/z/ecc/android/crypto/SecretKeyFactoryCommon.kt
@@ -1,0 +1,13 @@
+package cash.z.ecc.android.crypto
+
+expect class SecretKeyFactoryCommon {
+    fun generateSecret(pbeKeySpec: PBEKeySpecCommon): SecretKeyCommon
+
+    companion object {
+        fun getInstance(algorithm: String): SecretKeyFactoryCommon
+        fun getInstance(algorithm: String, provider: FallbackProvider): SecretKeyFactoryCommon
+    }
+}
+
+
+
