+++ lib/src/commonMain/kotlin/cash/z/ecc/android/random/SecureRandom.kt
@@ -1,0 +1,5 @@
+package cash.z.ecc.android.random
+
+expect class SecureRandom() {
+    fun nextBytes(bytes: ByteArray)
+}
