+++ lib/src/jvmMain/kotlin/cash/z/ecc/android/crypto/SecretKeyCommon.kt
@@ -1,0 +1,9 @@
+package cash.z.ecc.android.crypto
+
+import javax.crypto.SecretKey
+
+actual class SecretKeyCommon(generatedSecret: SecretKey) {
+
+    actual val encoded = generatedSecret.encoded
+
+}
