+++ lib/src/jvmMain/kotlin/cash/z/ecc/android/random/SecureRandom.kt
@@ -1,0 +1,8 @@
+package cash.z.ecc.android.random
+
+actual class SecureRandom {
+    val secureRandom = java.security.SecureRandom()
+
+    actual fun nextBytes(bytes: ByteArray) = secureRandom.nextBytes(bytes)
+
+}
