+++ lib/src/jvmTest/resources/data/BIP-0039-test-values.json
