+++ lib/src/nonJvmMain/kotlin/cash/z/ecc/android/crypto/FallbackProvider.kt
@@ -1,0 +1,3 @@
+package cash.z.ecc.android.crypto
+
+actual class FallbackProvider
