+++ lib/src/nonJvmMain/kotlin/cash/z/ecc/android/crypto/SecretKeyCommon.kt
@@ -1,0 +1,7 @@
+package cash.z.ecc.android.crypto
+
+actual class SecretKeyCommon {
+
+    actual val encoded : ByteArray = TODO()
+
+}
