+++ lib/src/nonJvmMain/kotlin/cash/z/ecc/android/crypto/SecretKeyFactoryCommon.kt
@@ -1,0 +1,18 @@
+package cash.z.ecc.android.crypto
+
+actual class SecretKeyFactoryCommon() {
+
+    actual fun generateSecret(pbeKeySpec: PBEKeySpecCommon): SecretKeyCommon {
+        TODO()
+    }
+
+    actual companion object{
+        actual fun getInstance(algorithm: String): SecretKeyFactoryCommon = TODO()
+
+
+        actual fun getInstance(algorithm: String, provider: FallbackProvider): SecretKeyFactoryCommon =  TODO()
+
+    }
+
+
+}
